% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{CabecerasGraph}
\alias{CabecerasGraph}
\title{Construir cabeceras HTTP con token Bearer}
\usage{
CabecerasGraph(...)
}
\value{Lista de cabeceras HTTP lista para usar con httr2.}
\description{
Construir cabeceras HTTP con token Bearer
}
