% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{CargarExcelDesdeOneDrive}
\alias{CargarExcelDesdeOneDrive}
\title{Cargar y abrir Excel desde OneDrive como workbook openxlsx2}
\usage{
CargarExcelDesdeOneDrive(...)
}
\arguments{
  \item{usuario}{UPN o alias del usuario.}
  \item{ruta}{Ruta relativa dentro del OneDrive.}
  \item{archivo}{Nombre del archivo.}
}
\value{Objeto `wb_workbook` de openxlsx2.}
\description{
Cargar y abrir Excel desde OneDrive como workbook openxlsx2
}
