% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{DescargarArchivoId}
\alias{DescargarArchivoId}
\title{Descargar archivo de OneDrive a un archivo temporal}
\usage{
DescargarArchivoId(...)
}
\arguments{
  \item{archivo_id}{ID del archivo en OneDrive.}
  \item{usuario}{UPN o alias del usuario.}
}
\value{Ruta al archivo temporal descargado.}
\description{
Descargar archivo de OneDrive a un archivo temporal
}
