% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{ListarContenidoCarpetaId}
\alias{ListarContenidoCarpetaId}
\title{Listar contenido de una carpeta por ID}
\usage{
ListarContenidoCarpetaId(...)
}
\arguments{
  \item{usuario}{UPN o alias del usuario.}
  \item{carpeta_id}{ID de la carpeta.}
}
\value{`data.frame` con metadatos de los elementos.}
\description{
Listar contenido de una carpeta por ID
}
