% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{ListarContenidoCarpetaNombre}
\alias{ListarContenidoCarpetaNombre}
\title{Listar contenido de una carpeta por nombre}
\usage{
ListarContenidoCarpetaNombre(...)
}
\arguments{
  \item{usuario}{UPN o alias del usuario.}
  \item{nombre_carpeta}{Nombre de la carpeta.}
}
\value{`data.frame` con metadatos de los elementos.}
\description{
Listar contenido de una carpeta por nombre
}
