% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{ListarHojasExcelOneDrive}
\alias{ListarHojasExcelOneDrive}
\title{Listar hojas de un Excel alojado en OneDrive}
\usage{
ListarHojasExcelOneDrive(...)
}
\arguments{
  \item{archivo_id}{ID del archivo Excel.}
  \item{usuario}{UPN o alias del usuario.}
}
\value{Vector de caracteres con nombres de hojas.}
\description{
Listar hojas de un Excel alojado en OneDrive
}
