% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{ListarTodoContenidoCarpeta}
\alias{ListarTodoContenidoCarpeta}
\title{Listar todos los archivos de una carpeta como tibble recursivo}
\usage{
ListarTodoContenidoCarpeta(...)
}
\arguments{
  \item{usuario}{UPN o alias del usuario.}
  \item{carpeta_id}{ID de la carpeta raiz.}
}
\value{`tibble` con todos los archivos incluyendo subcarpetas.}
\description{
Listar todos los archivos de una carpeta como tibble recursivo
}
