% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{ObtenerIdCarpeta}
\alias{ObtenerIdCarpeta}
\title{Obtener ID de una carpeta por nombre en el OneDrive de un usuario}
\usage{
ObtenerIdCarpeta(...)
}
\arguments{
  \item{usuario}{UPN o alias del usuario.}
  \item{nombre_carpeta}{Nombre de la carpeta.}
}
\value{Cadena con el folder ID.}
\description{
Obtener ID de una carpeta por nombre en el OneDrive de un usuario
}
