% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{ObtenerIdSite}
\alias{ObtenerIdSite}
\title{Obtener ID de sitio SharePoint}
\usage{
ObtenerIdSite(...)
}
\arguments{
  \item{hostname}{Hostname de SharePoint (ej. `"contoso.sharepoint.com"`).}
  \item{site_path}{Ruta del sitio (ej. `"sites/mi-sitio"`).}
}
\value{Cadena con el site ID.}
\description{
Obtener ID de sitio SharePoint
}
